+++ libs/feature/audit-builder/src/lib/audit-builder/property.component.ts
@@ -6,7 +6,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MatOptionModule } from '@angular/material/core';
 import { MatSelectModule } from '@angular/material/select';
 import { RxFor } from '@rx-angular/template/for';
-import { PropertyName, StepType } from './audit-builder.types';
+import { PropertyName, StepType } from '../schema/audit-builder.types';
 import { AuditBuilderService } from './audit-builder.service';
 import { INPUT_TYPE, PROPERTY_NAME } from '../schema/step-property.constants';
 import { MatIconModule } from '@angular/material/icon';
